backend/requirements.txt
python-dotenv==1.0.1
django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
langchain==0.3.7
langchain-google-genai==2.0.6
langchain-openai==0.2.10
langchain-community
openai>=0.27.0
PyPDF2==3.0.1